Added GradientContainer widget
A new StatelessWidget, GradientContainer, has been added. This widget builds a MaterialApp with a Scaffold that contains a Container. The container features a linear gradient background transitioning from purple at the top to green at the bottom. It also includes centered text displaying "Hello, World!" in white color with a font size of 28.